Add automated deployment with Ansible, improve camera service and sensor data
This commit adds an Ansible playbook for deploying the IoT project to multiple Raspberry Pi devices under the file deploy.yml. It copies the project files, configures the necessary packages, and sets up the appropriate crontab jobs for data sending and image capturing tasks. The related deploy.sh script is added to run this playbook.

Adjustments are also made on the camera settings with changes in param.py and camera_service.py files to manage camera brightness based on hostname.

In the send_sensor_data.py, the data sent to the client service is cast to an integer for consistency.

Lastly, a new ansible host file "inventory.ini" has been added to manage ansible hosts and their corresponding login details.